requirements.txt
discord.py
yt-dlp
apscheduler
aiohttp
